requirements.txt
selene==2.0.0rc2
pytest
allure-pytest
